refactor: move whatsapp connection into platform
